+++ src/bin/bat/app.rs
@@ -524,6 +524,7 @@ impl App {
             #[cfg(feature = "lessopen")]
             use_lessopen: self.matches.get_flag("lessopen"),
             set_terminal_title: self.matches.get_flag("set-terminal-title"),
+            markdown_tables: self.matches.get_flag("markdown-tables"),
             squeeze_lines: if self.matches.get_flag("squeeze-blank") {
                 Some(
                     self.matches